@@ -1,5 +1,6 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+// import ReactDOM from 'react-dom'; The new way to import createRoot:
+import { createRoot } from "react-dom/client";
 import { createStore, combineReducers, applyMiddleware } from 'redux';
 import { Provider } from 'react-redux';
 import thunkMiddleware from 'redux-thunk';
@@ -19,10 +20,17 @@ const rootReducers = combineReducers({requestRobots, searchRobots})
 
 const store = createStore(rootReducers, applyMiddleware(thunkMiddleware, logger))
 
-ReactDOM.render(
-  <Provider store={store}>
-    <App/>
-  </Provider>,
-  document.getElementById('root')
+const root = createRoot(document.getElementById('root'));
+root.render(
+<Provider store={store}>
+  <App/>
+</Provider>
 );
+
+// ReactDOM.render(
+//   <Provider store={store}>
+//     <App/>
+//   </Provider>,
+//   document.getElementById('root')
+// );
 registerServiceWorker();
